Only destory splunk server if enabled

+++ recipes/destroy_splunk.rb
@@ -22,42 +22,43 @@
 
 include_recipe 'delivery-cluster::_settings'
 
-# TODO: This should be moved out
-begin
-  with_chef_server chef_server_url,
-                   client_name: 'delivery',
-                   signing_key_filename: "#{cluster_data_dir}/delivery.pem"
+if splunk_enabled?
+  begin
+    with_chef_server chef_server_url,
+                     client_name: 'delivery',
+                     signing_key_filename: "#{cluster_data_dir}/delivery.pem"
 
-  directory "#{current_dir}/data_bags/vault" do
-    recursive true
-    action :delete
-  end
-
-  %W(
-    #{cluster_data_dir}/splunk.key
-    #{cluster_data_dir}/splunk.csr
-    #{cluster_data_dir}/splunk.crt
-  ).each do |f|
-    file f do
+    directory "#{current_dir}/data_bags/vault" do
+      recursive true
       action :delete
     end
-  end
 
-  # Delete Splunk ChefVault
-  execute 'Creating Splunk ChefVault' do
-    cwd current_dir
-    command 'knife data bag delete vault -y'
-    only_if 'knife data bag show vault'
-    action :run
-  end
+    %W(
+      #{cluster_data_dir}/splunk.key
+      #{cluster_data_dir}/splunk.csr
+      #{cluster_data_dir}/splunk.crt
+    ).each do |f|
+      file f do
+        action :delete
+      end
+    end
 
-  # Kill the machine
-  machine splunk_server_hostname do
-    action :destroy
-  end
+    # Delete Splunk ChefVault
+    execute 'Creating Splunk ChefVault' do
+      cwd current_dir
+      command 'knife data bag delete vault -y'
+      only_if 'knife data bag show vault'
+      action :run
+    end
 
-  # Delete the lock file
-  File.delete(splunk_lock_file)
-rescue StandardError => e
-  Chef::Log.warn("We can't proceed to destroy Splunk Sever: #{e.message}")
+    # Kill the machine
+    machine splunk_server_hostname do
+      action :destroy
+    end
+
+    # Delete the lock file
+    File.delete(splunk_lock_file)
+  rescue StandardError => e
+    Chef::Log.warn("We can't proceed to destroy Splunk Sever: #{e.message}")
+  end
 end
